+++ sycl/plugins/level_zero/CMakeLists.txt
@@ -111,6 +111,7 @@ add_sycl_plugin(level_zero
     "../unified_runtime/ur/adapters/level_zero/ur_level_zero_program.hpp"
     "../unified_runtime/ur/adapters/level_zero/ur_level_zero_queue.hpp"
     "../unified_runtime/ur/adapters/level_zero/ur_level_zero_sampler.hpp"
+    "../unified_runtime/ur/adapters/level_zero/ur_level_zero_usm.hpp"
     "../unified_runtime/ur/adapters/level_zero/ur_level_zero.cpp"
     "../unified_runtime/ur/adapters/level_zero/ur_level_zero_common.cpp"
     "../unified_runtime/ur/adapters/level_zero/ur_level_zero_context.cpp"
@@ -122,6 +123,7 @@ add_sycl_plugin(level_zero
     "../unified_runtime/ur/adapters/level_zero/ur_level_zero_program.cpp"
     "../unified_runtime/ur/adapters/level_zero/ur_level_zero_queue.cpp"
     "../unified_runtime/ur/adapters/level_zero/ur_level_zero_sampler.cpp"
+    "../unified_runtime/ur/adapters/level_zero/ur_level_zero_usm.cpp"
     # Following are the PI Level-Zero Plugin only codes.
     "pi_level_zero.cpp"
     "pi_level_zero.hpp"
